+++ mai-notices.php
@@ -4,7 +4,7 @@
  * Plugin Name:     Mai Notices
  * Plugin URI:      https://bizbudding.com/products/mai-notices/
  * Description:     Custom block for callout notices in your content.
- * Version:         1.1.2
+ * Version:         1.2.0
  *
  * Author:          BizBudding
  * Author URI:      https://bizbudding.com
@@ -92,7 +92,7 @@ final class Mai_Notices_Plugin {
 	private function setup_constants() {
 		// Plugin version.
 		if ( ! defined( 'MAI_NOTICES_VERSION' ) ) {
-			define( 'MAI_NOTICES_VERSION', '1.1.2' );
+			define( 'MAI_NOTICES_VERSION', '1.2.0' );
 		}
 
 		// Plugin Folder Path.
@@ -213,9 +213,9 @@ final class Mai_Notices_Plugin {
  *
  * @return object|Mai_Notices_Plugin The one true Mai_Notices_Plugin Instance.
  */
-function Mai_Notices_Plugin() {
+function mai_notices_plugin() {
 	return Mai_Notices_Plugin::instance();
 }
 
 // Get Mai_Notices_Plugin Running.
-Mai_Notices_Plugin();
+mai_notices_plugin();